backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==5.0.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.0
greenlet==3.4.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.3.0
langchain-groq==1.1.2
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-checkpoint-postgres==3.0.5
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.33
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
passlib==1.7.4
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic-core==2.46.3
pydantic-settings==2.9.1
pyjwt==2.12.1
python-dotenv==1.2.2
pyyaml==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlalchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.45.0
xxhash==3.6.0
zstandard==0.25.0
